Extract session recovery helpers
